Primitive array types in ArrayOps / ArrayBuilder / ImmutableArray
This is similar to the old implementation. Primitive array types are not
polymorphic on the JVM so we need to special-case them in order to be
able to efficiently wrap existing arrays as collections.

We also remove ArrayView in favor of ImmutableArray. With the new
equality semantics (where Views are never equal unless they are
the same reference) using a View as a default wrapper type for Array is
problematic. ImmutableArray is basically the same as ArrayView, except
it is a Seq (like WrappedArray in the old colections), so we get the
expected equality semantics and strict operations.